requirements.txt
pyaudio==0.2.13
tencentcloud-sdk-python==3.0.1035
pytest==7.4.3
black==23.11.0
flake8==6.1.0
python-dotenv==1.0.0
pydub==0.25.1
